requirements.txt
openai
streamlit
pandas
pdfplumber
python-docx
python-dotenv
